+++ dotNetInstallerToolsLib/ShellUtil.h
@@ -30,9 +30,9 @@ namespace DVLib
 	// detach a command, return process information
 	void DetachCmd(const std::wstring& cmd, LPPROCESS_INFORMATION lpi = NULL);
 	// run a command, return process information
-	void RunCmd(const std::wstring& cmd, LPPROCESS_INFORMATION lpi = NULL, int flags = 0, int nShow = SW_SHOWNORMAL);
+	void RunCmd(const std::wstring& cmd, LPPROCESS_INFORMATION lpi = NULL, int flags = 0, const std::wstring& working_directory = L"", int nShow = SW_SHOWNORMAL);
 	// execute a process, wait and return exit code
-	DWORD ExecCmd(const std::wstring& cmd, int nShow = SW_SHOWNORMAL);
+	DWORD ExecCmd(const std::wstring& cmd, const std::wstring& working_directory = L"", int nShow = SW_SHOWNORMAL);
 	// shell-execute a process
-	void ShellCmd(const std::wstring& cmd, int * rc = NULL, LPHANDLE lpProcessHandle = NULL, HWND hWnd = NULL, int nShow = SW_SHOWNORMAL);
+	void ShellCmd(const std::wstring& cmd, int * rc = NULL, LPHANDLE lpProcessHandle = NULL, HWND hWnd = NULL, const std::wstring& working_directory = L"", int nShow = SW_SHOWNORMAL);
 }